# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Concrete business API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type QuoteRequestStatus =
  (typeof QuoteRequestStatus)[keyof typeof QuoteRequestStatus];

export const QuoteRequestStatus = {
  pending: "pending",
  reviewed: "reviewed",
  converted: "converted",
  rejected: "rejected",
} as const;

export interface QuoteRequest {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  customerAddress?: string | null;
  projectDescription: string;
  concreteType: string;
  /** Volume in cubic yards */
  volumeEstimate: number;
  desiredDeliveryDate: string | null;
  status: QuoteRequestStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SubmitQuoteRequestBody {
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  customerAddress?: string;
  projectDescription: string;
  desiredDate?: string;
  notes?: string;
}

export type UpdateQuoteRequestStatusBodyStatus =
  (typeof UpdateQuoteRequestStatusBodyStatus)[keyof typeof UpdateQuoteRequestStatusBodyStatus];

export const UpdateQuoteRequestStatusBodyStatus = {
  pending: "pending",
  reviewed: "reviewed",
  converted: "converted",
  rejected: "rejected",
} as const;

export interface UpdateQuoteRequestStatusBody {
  status: UpdateQuoteRequestStatusBodyStatus;
  notes?: string;
}

export type QuoteStatus = (typeof QuoteStatus)[keyof typeof QuoteStatus];

export const QuoteStatus = {
  draft: "draft",
  sent: "sent",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface Quote {
  id: number;
  quoteRequestId?: number | null;
  customerName: string;
  customerEmail: string;
  status: QuoteStatus;
  totalAmount: number;
  expiryDate?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface QuoteLineItem {
  id: number;
  quoteId: number;
  description: string;
  unit: string;
  quantity: number;
  unitPrice: number;
  createdAt: string;
  updatedAt: string;
}

export type QuoteWithLineItems = Quote & {
  lineItems: QuoteLineItem[];
};

export interface DashboardStats {
  openRequestsCount: number;
  activeQuotesCount: number;
  unpaidInvoicesCount: number;
  recentRequests: QuoteRequest[];
  recentQuotes: Quote[];
}

export interface CreateQuoteRequestBody {
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  projectDescription: string;
  concreteType: string;
  volumeEstimate: number;
  desiredDeliveryDate: string;
}

export interface LineItem {
  id: number;
  description: string;
  unit: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface LineItemInput {
  id?: number;
  description: string;
  unit: string;
  quantity: number;
  unitPrice: number;
}

export type QuoteDetail = Quote & {
  lineItems: LineItem[];
};

export interface CreateQuoteBody {
  quoteRequestId?: number | null;
  customerName: string;
  customerEmail: string;
  expiryDate?: string | null;
  notes?: string | null;
  lineItems?: LineItemInput[];
}

export type UpdateQuoteStatusBodyStatus =
  (typeof UpdateQuoteStatusBodyStatus)[keyof typeof UpdateQuoteStatusBodyStatus];

export const UpdateQuoteStatusBodyStatus = {
  draft: "draft",
  sent: "sent",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface UpdateQuoteStatusBody {
  status: UpdateQuoteStatusBodyStatus;
}

export interface UpsertLineItemBody {
  description: string;
  unit: string;
  quantity: number;
  unitPrice: number;
}

export type UpdateQuoteBodyStatus =
  (typeof UpdateQuoteBodyStatus)[keyof typeof UpdateQuoteBodyStatus];

export const UpdateQuoteBodyStatus = {
  draft: "draft",
  sent: "sent",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface UpdateQuoteBody {
  status?: UpdateQuoteBodyStatus;
  notes?: string | null;
  expiryDate?: string | null;
  lineItems?: LineItemInput[];
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface Invoice {
  id: number;
  quoteId: number;
  customerName: string;
  customerEmail: string;
  totalAmount: number;
  status: InvoiceStatus;
  dueDate?: string | null;
  paidAt?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type UpdateInvoiceStatusBodyStatus =
  (typeof UpdateInvoiceStatusBodyStatus)[keyof typeof UpdateInvoiceStatusBodyStatus];

export const UpdateInvoiceStatusBodyStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface UpdateInvoiceStatusBody {
  status: UpdateInvoiceStatusBodyStatus;
}