# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Concrete business API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quote-requests:
    get:
      operationId: listQuoteRequests
      tags: [quoteRequests]
      summary: List all quote requests (staff)
      responses:
        "200":
          description: List of quote requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteRequest"
    post:
      operationId: submitQuoteRequest
      tags: [quoteRequests]
      summary: Submit a quote request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuoteRequestBody"
      responses:
        "201":
          description: Quote request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quote-requests/{id}:
    get:
      operationId: getQuoteRequest
      tags: [quoteRequests]
      summary: Get a single quote request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /invoices/from-quote/{quoteId}:
    post:
      operationId: generateInvoiceFromQuote
      tags: [invoices]
      summary: Generate an invoice from an accepted quote
      parameters:
        - name: quoteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Quote is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get a single invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{id}/status:
    patch:
      operationId: updateInvoiceStatus
      tags: [invoices]
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceStatusBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark an invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    QuoteRequest:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        customerAddress:
          type: string
          nullable: true
        projectDescription:
          type: string
        concreteType:
          type: string
        volumeEstimate:
          type: number
          description: Volume in cubic yards
        desiredDeliveryDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, reviewed, converted, rejected]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - projectDescription
        - concreteType
        - volumeEstimate
        - status
        - createdAt
        - updatedAt
    SubmitQuoteRequestBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        projectDescription:
          type: string
        concreteType:
          type: string
        volumeEstimate:
          type: number
        desiredDeliveryDate:
          type: string
        notes:
          type: string
      required:
        - customerName
        - customerEmail
        - projectDescription
        - concreteType
        - volumeEstimate
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [unpaid, paid]
        dueDate:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - quoteId
        - customerName
        - customerEmail
        - totalAmount
        - status
        - createdAt
    UpdateInvoiceStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [unpaid, paid]
      required:
        - status