# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Concrete business API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceStatus } from "./invoiceStatus";

export interface Invoice {
  id: number;
  quoteId: number;
  customerName: string;
  customerEmail: string;
  totalAmount: number;
  status: InvoiceStatus;
  dueDate?: string | null;
  paidAt?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}