# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Concrete business API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteRequestStatus } from "./quoteRequestStatus";

export interface QuoteRequest {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  customerAddress?: string | null;
  projectDescription: string;
  concreteType: string;
  /** Volume in cubic yards */
  volumeEstimate: number;
  desiredDeliveryDate?: string | null;
  status: QuoteRequestStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}